Support for single-file publish for .NET 7.0

+++ release/dotnet/7.0.sh
@@ -8,12 +8,14 @@ if [ -z "${VERSION}" ]; then
 fi
 
 APP_NAME=$(git remote -v | tail -1 | sed 's|.*/\([^/]*\)\.git.*|\1|')
+MAIN_PROJECT=$(pwd)
 
 if [ -f *.sln ]; then
     MAIN_PROJECT=$(ls *.sln | head -n 1 | xargs cat | grep "^Project" | head -n 1 | awk -F"=" '{print $2}' | awk -F"," '{print $1}' | sed -e 's/\"*//g' -e 's/\s*//g')
     cd "${MAIN_PROJECT}"
 fi
 
+BINARY_FILE_LABEL=$(cat "${MAIN_PROJECT}"/*.csproj | grep "RootNamespace" | sed 's/[^>]*>\([^<]*\).*/\1/g')
 PROJECT_SOURCE_DIR="$(pwd)"
 BIN_RELEASE_DIR="${PROJECT_SOURCE_DIR}/bin/Release"
 PUBLISH_DIR="${BIN_RELEASE_DIR}/.publish-script-output"
@@ -21,15 +23,31 @@ PUBLISH_DIR="${BIN_RELEASE_DIR}/.publish-script-output"
 function package {
     local ARCH="${1}"
     local OUTPUT_DIR="${PUBLISH_DIR}/${ARCH}"
-    local OUTPUT_FILE="${BIN_RELEASE_DIR}/${APP_NAME}_${VERSION}_${ARCH}.zip"
+    local OUTPUT_FILE=""
+    local FILES_COUNT=0
 
-    echo "Packaging \"${OUTPUT_DIR}\" to \"${OUTPUT_FILE}\""
+    rm "${OUTPUT_DIR}"/*.pdb
+    FILES_COUNT=$(ls -1q "${OUTPUT_DIR}" | wc -l)
 
-    [ -f "${OUTPUT_FILE}" ] && rm "${OUTPUT_FILE}"
+    if [ ${FILES_COUNT} -eq 1 ]; then
+        BINARY_FILE=$(ls "${OUTPUT_DIR}"/*)
+        BINARY_FILE_NAME=$(basename "${BINARY_FILE}")
+        OUTPUT_FILE=$(sed 's/'"${BINARY_FILE_LABEL}"'/'"${APP_NAME}"'_'"${VERSION}"'_'"${ARCH}"'/g' <<< "${BINARY_FILE_NAME}")
+        OUTPUT_FILE="${BIN_RELEASE_DIR}/${OUTPUT_FILE}"
 
-    cd "${OUTPUT_DIR}" || exit
-    zip -q -9 -r "${OUTPUT_FILE}" .
-    cd "${PROJECT_SOURCE_DIR}" || exit
+        echo "Copying \"${BINARY_FILE_NAME}\" to \"${OUTPUT_FILE}\""
+
+        cp "${BINARY_FILE}" "${OUTPUT_FILE}"
+    else
+        OUTPUT_FILE="${BIN_RELEASE_DIR}/${APP_NAME}_${VERSION}_${ARCH}.zip"
+
+        echo "Packaging \"${OUTPUT_DIR}\" as \"${OUTPUT_FILE}\""
+        [ -f "${OUTPUT_FILE}" ] && rm "${OUTPUT_FILE}"
+
+        cd "${OUTPUT_DIR}" || exit
+        zip -q -9 -r "${OUTPUT_FILE}" .
+        cd "${PROJECT_SOURCE_DIR}" || exit
+    fi
 }
 
 function dotnet-pub {
@@ -75,4 +93,3 @@ build-release win-arm64
 build-release win-x64
 
 cleanup
-
